add servfail stats, make cachehits/misses unsigend
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns/pdns@605 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ recursor_cache.hh
@@ -29,7 +29,7 @@ public:
   int get(time_t, const string &qname, const QType& qt, set<DNSResourceRecord>* res);
   void replace(const string &qname, const QType& qt,  const set<DNSResourceRecord>& content);
   void doPrune(void);
-  int cacheHits, cacheMisses;
+  uint64_t cacheHits, cacheMisses;
 
 private:
   struct StoredRecord
